update id type

+++ goyotashi/Developing/TestData.swift
@@ -89,7 +89,7 @@ struct TestData {
 
     static func groupSummary() -> GroupSummary {
         return GroupSummary(
-            groupId: randomInt(),
+            groupId: randomInt64(),
             groupName: "CAMPHOR-",
             groupDescription: "CAMPHOR-カンファーは京都のIT系学生コミュニティです。エンジニアリング・デザイン・プロダクト開発などへの関心を共通点とする、様々なバックグラウンドを持つ学生が集まっています。",
             restaurantCount: Int.random(in: 2 ..< 16),
